+++ src/ItunesSearchAPIProvider.php
@@ -13,7 +13,11 @@ class ItunesSearchAPIProvider extends ServiceProvider
      */
     public function boot()
     {
-        //
+        $this->publishes([
+            __DIR__.'/config/laravel-itunes-search-api.php' => config_path('laravel-itunes-search-api.php'),
+        ], 'config');
+
+        $this->mergeConfigFrom(__DIR__.'/config/laravel-itunes-search-api.php', 'laravel-itunes-search-api');
     }
 
     /**