+++ src/config/laravel-itunes-search-api.php
@@ -1,0 +1,9 @@
+<?php
+
+return [
+
+    /*
+     * By default, the cache is set to 60 minutes
+     */
+    'cache' => 60,
+];