+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All Notable changes to `laravel-itunes-store-api` will be documented in this fil
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
+## Unreleased - YYYY-MM-DD
+### Added
+- Config file with default cache time
+
+### Deprecated
+### Fixed
+### Removed
+### Security
+
 ## 0.0.1 - 2016-10-12
 
 ### Added